requirements.txt

aiogram==3.3.0
python-dotenv>=0.21.0
httpx>=0.25.0
apscheduler>=3.10.0
asyncpg>=0.27.0
reportlab>=3.6.0
